requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
python-dotenv
fastapi
uvicorn[standard]
pydantic
pydantic-settings
pytest
pytest-asyncio
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator
pyjwt
